Accept Integer in take_first_char for 32-bit Julia. (#168)
n_char_per_name is Int (Int32 on i686); the Int64-only method never matched.

+++ src/print_human_readable.jl
@@ -6,7 +6,7 @@ function add_separator!(first_element, toprint, level)
     return false
 end
 
-function take_first_char(s::Union{String, SubString{String}}, N::Int64)
+function take_first_char(s::Union{String, SubString{String}}, N::Integer)
     return s[1:min(length(s), N)]
 end
